porting 2022 sep 27 1241
