+++ src/font_atlas.rs
@@ -3,12 +3,13 @@
 use std::ptr::{null, null_mut};
 use libc::{c_char, c_float, c_int, c_uchar, c_uint, c_ushort, c_void, size_t};
 use crate::color::IM_COL32;
-use crate::draw::{Decode85, ImFontAtlasGetBuilderForStbTruetype, UnpackAccumulativeOffsetsIntoRanges};
+use crate::draw::{Decode85, UnpackAccumulativeOffsetsIntoRanges};
 use crate::file_ops::ImFileLoadToMemory;
 use crate::font::ImFont;
 use crate::font_atlas_custom_rect::ImFontAtlasCustomRect;
 use crate::font_atlas_default_tex_data::{FONT_ATLAS_DEFAULT_TEX_CURSOR_DATA, FONT_ATLAS_DEFAULT_TEX_DATA_W};
 use crate::font_atlas_flags::{ImFontAtlasFlags, ImFontAtlasFlags_NoMouseCursors};
+use crate::font_atlas_ops::ImFontAtlasGetBuilderForStbTruetype;
 use crate::font_config::ImFontConfig;
 use crate::font_glyph::ImFontGlyph;
 use crate::mouse_cursor::{ImGuiMouseCursor, ImGuiMouseCursor_COUNT, ImGuiMouseCursor_None};
